requirements.txt
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
flask==3.0.0
flask-cors==4.0.0
websocket-client==1.7.0
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
upstox-python-sdk
gunicorn==21.2.0
psycopg2-binary==2.9.9
schedule==1.2.0
